Backend/requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
python-dotenv
google-generativeai
PyMuPDF
